refactor: move runtime_promotion_signals into astra-evolution (446 LOC)
The module defines signal types for the evolution promotion gate, making
astra-evolution its natural home. Its only non-evolution dep is
TurnEvaluation (now in astra-turn-core), which astra-evolution already imports.

- astra-evolution/src/promotion_signals.rs (new)
- runtime/src/runtime_promotion_signals.rs → 6-line re-export stub
- astra-evolution gains astra-core + astra-services deps for ConfidenceInterval, ValueInterval

Stacks on PRs #124-#126.
